Stop codegen from exploding by moving kindsys Resource out
Generating OpenAPI definitions for googlesheets:v1 at github.com/grafana/google-sheets-datasource/pkg/client/openapi
E0920 16:47:27.454613   12524 openapi.go:321] Error when generating: migrator, migrator func(ctx context.Context, obj github.com/grafana/kindsys.Resource, targetVersion string) (github.com/grafana/kindsys.Resource, error)
2023/09/20 16:47:27 OpenAPI code generation error: Failed executing generator: some packages had errors:
failed to generate default in github.com/grafana/google-sheets-datasource/pkg/apis/googlesheets/v1.rawSheetsKind: migrator: not sure how to enforce default for Func
